# Conflicts:
#	src/main/java/com/Database/Persistence.java
#	src/main/java/com/Database/storedProcedures.java
#	src/main/java/com/IceHockeyLeague/LeagueManager/League/League.java
#	src/main/java/com/IceHockeyLeague/LeagueManager/Team/ITeam.java
#	src/main/java/com/IceHockeyLeague/LeagueManager/Team/Team.java

+++ src/main/java/com/IceHockeyLeague/LeagueManager/Player/IPlayerStats.java
@@ -16,7 +16,7 @@ public interface IPlayerStats {
     int getSaving();
     void setSaving(int saving);
 
-    void setStrength(int strength);
-    int getStrength();
-    int calculateStrength();
+    void setStrength(float strength);
+    float getStrength();
+    float calculateStrength();
 }
